requirements.txt
# Scalping Bot V2 - Dependencies
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
google-genai>=0.5.0
pydantic>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
